// demos/customer-feedback-assistant/src/generated/models/Crfaa_feedbacksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crfaa_feedbackscrfaa_channel = {
  1: 'Web',
  2: 'Email',
  3: 'Phone',
  4: 'In_Person',
  5: 'Social'
} as const;
export type Crfaa_feedbackscrfaa_channel = keyof typeof Crfaa_feedbackscrfaa_channel;
export const Crfaa_feedbackscrfaa_sentiment = {
  1: 'Positive',
  2: 'Neutral',
  3: 'Negative'
} as const;
export type Crfaa_feedbackscrfaa_sentiment = keyof typeof Crfaa_feedbackscrfaa_sentiment;
export const Crfaa_feedbackscrfaa_status = {
  1: 'New',
  2: 'InReview',
  3: 'Actioned',
  4: 'Closed'
} as const;
export type Crfaa_feedbackscrfaa_status = keyof typeof Crfaa_feedbackscrfaa_status;
export const Crfaa_feedbackscrfaa_theme = {
  1: 'Product',
  2: 'Service',
  3: 'Pricing',
  4: 'Support',
  5: 'Documentation',
  6: 'Other'
} as const;
export type Crfaa_feedbackscrfaa_theme = keyof typeof Crfaa_feedbackscrfaa_theme;
export const Crfaa_feedbacksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crfaa_feedbacksstatecode = keyof typeof Crfaa_feedbacksstatecode;
export const Crfaa_feedbacksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crfaa_feedbacksstatuscode = keyof typeof Crfaa_feedbacksstatuscode;

export interface Crfaa_feedbacksBase {
  crfaa_channel?: Crfaa_feedbackscrfaa_channel;
  crfaa_customeremail?: string;
  crfaa_customername?: string;
  crfaa_feedbackid: string;
  crfaa_feedbacktext?: string;
  crfaa_name?: string;
  crfaa_rating?: number;
  crfaa_sentiment?: Crfaa_feedbackscrfaa_sentiment;
  crfaa_status?: Crfaa_feedbackscrfaa_status;
  crfaa_submittedon?: string;
  crfaa_suggestedresponse?: string;
  crfaa_theme?: Crfaa_feedbackscrfaa_theme;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crfaa_feedbacksstatecode;
  statuscode?: Crfaa_feedbacksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crfaa_feedbacks extends Crfaa_feedbacksBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crfaa_channelname?: string;
  crfaa_sentimentname?: string;
  crfaa_statusname?: string;
  crfaa_themename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}