// demos/customer-feedback-assistant/src/generated/models/Crfaa_actionitemsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crfaa_actionitemscrfaa_priority = {
  1: 'Low',
  2: 'Medium',
  3: 'High'
} as const;
export type Crfaa_actionitemscrfaa_priority = keyof typeof Crfaa_actionitemscrfaa_priority;
export const Crfaa_actionitemscrfaa_status = {
  1: 'Open',
  2: 'InProgress',
  3: 'Done'
} as const;
export type Crfaa_actionitemscrfaa_status = keyof typeof Crfaa_actionitemscrfaa_status;
export const Crfaa_actionitemsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crfaa_actionitemsstatecode = keyof typeof Crfaa_actionitemsstatecode;
export const Crfaa_actionitemsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crfaa_actionitemsstatuscode = keyof typeof Crfaa_actionitemsstatuscode;

export interface Crfaa_actionitemsBase {
  crfaa_actionitemid: string;
  crfaa_assignedto?: string;
  crfaa_description?: string;
  crfaa_duedate?: string;
  "crfaa_feedbackid@odata.bind"?: string;
  crfaa_name?: string;
  crfaa_priority?: Crfaa_actionitemscrfaa_priority;
  crfaa_status?: Crfaa_actionitemscrfaa_status;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crfaa_actionitemsstatecode;
  statuscode?: Crfaa_actionitemsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crfaa_actionitems extends Crfaa_actionitemsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crfaa_feedbackidname?: string;
  crfaa_priorityname?: string;
  crfaa_statusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  crfaa_feedbackid?: object;
  _crfaa_feedbackid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}